Remove helper generation
