+++ homeassistant/const.py
@@ -6,7 +6,7 @@ from enum import StrEnum
 from functools import partial
 from typing import TYPE_CHECKING, Final
 
-from .generated.platform_entity import EntityPlatform
+from .generated.entity_platforms import EntityPlatforms
 from .helpers.deprecation import (
     DeprecatedConstant,
     DeprecatedConstantEnum,
@@ -38,7 +38,7 @@ REQUIRED_NEXT_PYTHON_HA_RELEASE: Final = ""
 PLATFORM_FORMAT: Final = "{platform}.{domain}"
 
 # Type alias to avoid 1000 MyPy errors
-Platform = EntityPlatform
+Platform = EntityPlatforms
 
 BASE_PLATFORMS: Final = {platform.value for platform in Platform}
