+++ script/hassfest/__main__.py
@@ -19,7 +19,7 @@ from . import (
     dhcp,
     docker,
     icons,
-    integration_enums,
+    integration_info,
     json,
     manifest,
     metadata,
@@ -45,7 +45,7 @@ INTEGRATION_PLUGINS = [
     dependencies,
     dhcp,
     icons,
-    integration_enums,
+    integration_info,
     json,
     manifest,
     mqtt,
